Revert "fix(frontend): account-only web QR sign-in and separate login approva…"
This reverts commit 157b5467e6b0b58dc96f9b75e6990a43e16e11b9.
