+++ CLAUDE.md
@@ -171,13 +171,13 @@ First-party RFC 8628-style login: plain `nyxid login` and `--device` display a s
 
 - **Storage and rollout**: legacy `auth_device_codes` remains account-only. New `/auth/device/v2/{request,poll,poll-web}` uses isolated `auth_device_codes_v2`, `nyx_adc2_` poll secrets and `2-XXXX-XXXX` human codes. Shared preview/decisions choose storage by code version. Never retry v2 through legacy polling. Both collections store HMACs of codes, sanitized context and encrypted deliveries; raw codes and token plaintext never persist. Terminal TTL uses `purge_at` only after cleanup; an indexed expiry sweep retains unfinished revocations. Old replicas and legacy TTL indexes cannot consume/delete v2 grants.
 - **Routes**: `POST /auth/device/request`, `/poll`, and `/preview` are public (no auth). `/poll-web` is also public but is a first-party browser-only delivery surface: it shares the `/poll` limiter, atomically claims the same delivery, revokes the approve-minted JWT session, and creates a fresh `nyx_session` cookie session without returning tokens. Preview is non-mutating and returns untrusted `client_label`/`client_user_agent` plus server-observed `client_ip`, Z-suffix timestamps, and status. `POST /auth/device/approve` and `/deny` are human-only: JWT session user; API keys, service accounts, delegated tokens, and relay tokens are rejected before handler execution. Approve and deny race atomically on `status = pending`; deny mints no tokens and the next poll returns 11204. All routes are rate-limited per IP; decisions also share a per-user limit.
-- **Grant decisions and delivery**: only first-party human sessions/access JWTs without an OAuth client ID can decide grants. Restricted selection uses `/auth/device/approve-agent-key`, so an old replica cannot silently issue account access. Issuance and approval commit together. Decrypt before atomically claiming `approved` -> `delivered`, retaining retryability on KMS failures. Account browser delivery atomically revokes approval bearer credentials and creates the cookie session. `/auth/device/v2/poll-web` can still return a one-time CLI handoff for a restricted v2 approval (handoff redemption and direct polling have one winner), but no first-party page calls it: browser sign-in is always an account session. Abandonment revokes only its session/child, preserving reused parent configuration and other consumers.
+- **Grant decisions and delivery**: only first-party human sessions/access JWTs without an OAuth client ID can decide grants. Restricted selection uses `/auth/device/approve-agent-key`, so an old replica cannot silently issue account access. Issuance and approval commit together. Decrypt before atomically claiming `approved` -> `delivered`, retaining retryability on KMS failures. Account browser delivery atomically revokes approval bearer credentials and creates the cookie session. Restricted browser delivery returns a one-time CLI handoff without cookies; handoff redemption and direct polling have one winner. Abandonment revokes only its session/child, preserving reused parent configuration and other consumers.
 - **Rate limiters**: `auth_device_{request,poll,approve,approve_per_user,preview}_limiter` (`/poll-web` shares `poll`)
 - **CLI output**: prints `user_code` plus the bare `verification_uri` -- never `verification_uri_complete`, which would put the code in the URL and defeat manual-entry anti-phishing. On stdin+stderr TTYs: prompts `Open in your browser? [Y/n]` and opens via `crate::browser::open_browser`, falling through with a "paste it manually" hint on failure. Non-TTY (CI/piped) skips the prompt and polls immediately.
 - **CLI dispatch**: plain blocking login uses selectable device v2 and offers browser approval. `NYXID_LOGIN_NO_DEVICE_FALLBACK=1` preserves the legacy browser-callback strategy and its established browser-open fallback. Explicit device/Agent Key/no-wait modes use the exchange. `--no-wait --output json` never waits or opens a browser; `login resume <REQUEST_ID> [--once]` uses a mode-0600 local handle bound to profile/destination, persists backoff/deadlines and returns distinct stable JSON errors. Poll secrets never appear in output.
 - **Verification page** (`/login/device`): public preview with explicit computer/phone choice. Code input starts empty; query parameters are stripped and ignored. Account approval and key permission confirmation require first-party human authentication. No decision occurs on typing, mount or reconnect. Legacy codes only offer account access. Phone preview, including polling for the requesting computer, leaves browser account cookies/storage untouched. Login-grant routes suppress telemetry even after SPA navigation from an opted-in page.
 - **Mobile approval**: `CameraView` scans only QR codes whose shape is `https://<frontend>/login/device?user_code=...` or `nyxid://login/device?user_code=...`. The app extracts and normalizes only `user_code`; every request uses its configured API base URL. Deep links prefill without making a request. Preview and approve/reject remain separate explicit actions, share the >= 750 ms throttle, and never auto-approve.
-- **Web login**: the unauthenticated login page can start the browser delivery flow with an explicit click, showing both the QR payload and manual code; it never starts on mount or auto-regenerates codes. It uses only the legacy `/auth/device/request` and `/auth/device/poll-web` routes, so a browser signing itself in always requests and receives a full account session and its 8-character code stays scannable by every shipped mobile build. Restricted Agent Keys are requested only by terminal-originated flows (`nyxid login --agent-key`); never point the login page at `/auth/device/v2/*`.
+- **Web login**: the unauthenticated login page can start the browser delivery flow with an explicit click, showing both the QR payload and manual code; it never starts on mount or auto-regenerates codes.
 - **One-time codes**: web/mobile settings mint through `/auth/login-code`; public `/redeem` atomically claims and issues either grant. Issuer-only status, cancellation and revocation expose no secrets. Five-minute codes are limited by source and owner and stored only as HMACs. Browser handoff extends the initial 60-second delivery window to approval + 300 seconds; repeated handoffs never extend that fixed deadline. First-party JWT `sid` checks live session state during protected access and refresh; issuer revocation cascades to MCP and invalidates refreshed access too.
 - **Local identity lifecycle**: login replacement, refresh snapshots and logout use a profile generation under the refresh lock. An in-flight request cannot retry another login's token, including a replacement at the same destination. Bounded logout clears only the matching generation. Explicit credentials do not require profile storage or fall back to another identity.
