+++ docs/site/cli/getting-started/authenticate.md
@@ -76,9 +76,10 @@ form `--code XXXX-XXXX` supports automation but may enter shell history.
 Codes are consumed once. The issuing screen shows redemption and requester
 context; Cancel stops a pending code, while Revoke invalidates its delivered
 session or child credential. Neither the code nor a poll secret belongs in URLs
-or browser storage. The web login page's QR sign-in always produces an account
-session for that browser; use `nyxid login --agent-key` or a one-time code when
-a machine should receive a restricted Agent Key instead.
+or browser storage. A browser-owned QR login approved as restricted returns a
+one-time CLI handoff code instead of an account cookie.
+Creating that handoff extends the initial 60-second delivery window to five
+minutes after approval. Repeated handoff requests never extend the fixed deadline.
 
 ### Non-interactive credentials
 
@@ -104,7 +105,7 @@ Approval issues a new login credential bound to the selected key. Selecting an e
 
 Identity output includes the credential's hostname/profile label. `status` then lists the account, AI services, API keys, and nodes; sections denied by the key's scope display "unavailable with this key's scope". JSON output includes an `auth` object and uses `null` for unavailable sections. A missing local credential prompts reauthorization for that profile; a server rejection still fails the command.
 
-`nyxid logout --profile home-agent` attempts bounded server revocation and clears the matching local login, reporting whether revocation succeeded. A newer concurrent login is preserved. In the web console, open the key's **Login credentials** section to revoke a specific CLI login. Revoking or rotating the key invalidates every credential issued under it. Revocation and expiry take effect on subsequent authenticated requests. Abandoned approvals expire after a 60-second delivery window and their credentials are revoked automatically. A parent created for an abandoned request remains as key configuration with no disclosed primary secret; cleanup revokes only that request's child so another approved consumer remains usable.
+`nyxid logout --profile home-agent` attempts bounded server revocation and clears the matching local login, reporting whether revocation succeeded. A newer concurrent login is preserved. In the web console, open the key's **Login credentials** section to revoke a specific CLI login. Revoking or rotating the key invalidates every credential issued under it. Revocation and expiry take effect on subsequent authenticated requests. Abandoned approvals expire after a 60-second delivery window and their credentials are revoked automatically; a browser-owned device handoff uses the fixed five-minute deadline described above. A parent created for an abandoned request remains as key configuration with no disclosed primary secret; cleanup revokes only that request's child so another approved consumer remains usable.
 
 `--agent-key` cannot be combined with `--device` or `--password`. It supports headless polling, including a human authorizing a waiting CI job; unattended jobs should normally use a pre-issued credential through the existing environment-variable options.
