+++ frontend/src/hooks/use-auth-device.test.tsx
@@ -193,7 +193,32 @@ const requestResponse = {
 };
 
 describe("useWebAuthDeviceLogin", () => {
-  it.each(["account", "error"])("ignores an old %s poll after close and restart", async (outcome) => {
+  const restrictedDelivery = {
+    ok: false,
+    auth_kind: "agent_key",
+    login_code: { request_id: "handoff", code: "JKLM-NPQR", expires_at: "2026-08-20T10:00:10Z" },
+  };
+
+  it("clears the restricted handoff at its own expiry and permits regeneration", async () => {
+    vi.useFakeTimers();
+    vi.setSystemTime(new Date("2026-08-20T10:00:00Z"));
+    mockPost.mockResolvedValueOnce(requestResponse).mockResolvedValueOnce(restrictedDelivery);
+    const { result } = renderHook(() => useWebAuthDeviceLogin());
+    await act(async () => { result.current.start(); });
+    await act(async () => { await vi.advanceTimersByTimeAsync(5_000); });
+    expect(result.current.phase).toBe("restricted");
+    expect(result.current.loginCode?.code).toBe("JKLM-NPQR");
+    expect(mockCheckAuth).not.toHaveBeenCalled();
+    await act(async () => { await vi.advanceTimersByTimeAsync(5_000); });
+    expect(result.current.phase).toBe("expired");
+    expect(result.current.loginCode).toBeNull();
+    mockPost.mockResolvedValueOnce({ ...requestResponse, device_code: "nyx_adc_new" });
+    await act(async () => { result.current.generateNew(); });
+    expect(result.current.phase).toBe("pending");
+    expect(result.current.request?.device_code).toBe("nyx_adc_new");
+  });
+
+  it.each(["restricted", "account", "error"])("ignores an old %s poll after close and restart", async (outcome) => {
     vi.useFakeTimers();
     vi.setSystemTime(new Date("2026-08-20T10:00:00Z"));
     let finish!: (value: unknown) => void;
@@ -208,27 +233,12 @@ describe("useWebAuthDeviceLogin", () => {
     await act(async () => { result.current.start(); });
     await act(async () => {
       if (outcome === "error") fail({ errorCode: 11204 });
-      else finish({ ok: true, auth_kind: "account_session" });
+      else finish(outcome === "restricted" ? restrictedDelivery : { ok: true });
     });
     expect(result.current.phase).toBe("pending");
     expect(result.current.request?.device_code).toBe("nyx_adc_new");
+    expect(result.current.loginCode).toBeNull();
     expect(mockCheckAuth).not.toHaveBeenCalled();
-  });
-
-  it("rejects a restricted handoff delivery instead of showing a code", async () => {
-    vi.useFakeTimers();
-    vi.setSystemTime(new Date("2026-08-20T10:00:00Z"));
-    mockPost.mockResolvedValueOnce(requestResponse).mockResolvedValueOnce({
-      ok: false,
-      auth_kind: "agent_key",
-      login_code: { request_id: "handoff", code: "JKLM-NPQR", expires_at: "2026-08-20T10:05:00Z" },
-    });
-    const { result } = renderHook(() => useWebAuthDeviceLogin());
-    await act(async () => { result.current.start(); });
-    await act(async () => { await vi.advanceTimersByTimeAsync(5_000); });
-    expect(result.current.phase).toBe("pending");
-    expect(mockCheckAuth).not.toHaveBeenCalled();
-    expect(JSON.stringify(result.current)).not.toContain("JKLM-NPQR");
   });
 
   it("does not finish a replaced request after checkAuth resolves", async () => {
@@ -261,7 +271,7 @@ describe("useWebAuthDeviceLogin", () => {
 
     await waitFor(() => expect(result.current.phase).toBe("pending"));
     expect(mockPost).toHaveBeenCalledWith(
-      "/auth/device/request",
+      "/auth/device/v2/request",
       expect.objectContaining({
         client_label: expect.stringMatching(/ on /),
         client_user_agent: expect.any(String),
@@ -304,9 +314,6 @@ describe("useWebAuthDeviceLogin", () => {
       await vi.advanceTimersByTimeAsync(1);
     });
     expect(mockPost).toHaveBeenCalledTimes(2);
-    expect(mockPost).toHaveBeenLastCalledWith("/auth/device/poll-web", {
-      device_code: "nyx_adc_test",
-    });
 
     await act(async () => {
       await vi.advanceTimersByTimeAsync(9_999);
@@ -368,7 +375,7 @@ describe("useWebAuthDeviceLogin", () => {
 
     mockPost
       .mockRejectedValueOnce(new Error("temporary network failure"))
-      .mockResolvedValueOnce({ ok: true, auth_kind: "account_session" });
+      .mockResolvedValueOnce({ ok: true });
     await act(async () => {
       await vi.advanceTimersByTimeAsync(5_000);
     });
