+++ frontend/src/router.tsx
@@ -77,7 +77,6 @@ import {
   CliPairPage,
   LoginDevicePage,
   LoginAgentKeyPage,
-  LoginCodePage,
   ConnectLinkPage,
   ConnectLinkReturnPage,
   OAuthLaunchingPage,
@@ -277,7 +276,7 @@ const loginCodeRoute = createRoute({
   path: "/login/code",
   getParentRoute: () => rootRoute,
   validateSearch: (): Record<string, never> => ({}),
-  component: LoginCodePage,
+  component: () => <LoginAgentKeyPage mint />,
 });
 
 const connectLinkRoute = createRoute({
